requirements.txt
fastapi
uvicorn
python-jose
bcrypt
openpyxl
python-multipart
jinja2
PyQt5
